+++ src/main/java/de/samply/app/ProjectManagerApplication.java
@@ -14,8 +14,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 public class ProjectManagerApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(ProjectManagerApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(ProjectManagerApplication.class, args);
+    }
 
 }